feat: add cloud_firestore plugin to Flutter project
This commit includes the addition of the cloud_firestore plugin to the Flutter project. The plugin is registered in the GeneratedPluginRegistrant.swift file for macOS, the generated_plugin_registrant.cc file for Windows, and added to the dependencies in the pubspec.yaml file. This plugin will allow the application to interact with the Firestore database.